# src/schemas/openapi.yaml
openapi: 3.0.0
info:
  version: 1.0.0
  title: "We Sell Houses API — 304CEM Assignment"
  description: "A web application that allows users (estate agents) to list houses for sale to the general public."
  contact:
    name: Bruno Correia
    email: olveirb@coventry.ac.uk
servers:
  - url: https://photo-shrink-3000.codio-box.uk/api/
    description: "Development"
...
paths:
  /users/:
    post:
      tags:
        - Users
      summary: creates a new user
      description: receives a \'User\' object, validates it & uses it to create a new user
      parameters:
        - name: body
          in: query
          description: \'User\' object
          schema:
            type: object
      responses:
        "201":
          description: user created successfully

    get:
      tags:
        - Users
      summary: fetches all users
      description: returns a list of all users in the database
      responses:
        "200":
          description: user fetched successfully
  /users/{id}:
    get:
      tags:
        - Users
      summary: fetches a user
      description: receives an id and checks the database for a matching record which it then returns
      parameters:
        - name: id
          in: path
          description: user ID
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: user fetched successfully
    put:
      tags:
        - Users
      summary: updates a user
      description: updates the record matching the id parameter
      parameters:
        - name: id
          in: path
          description: user ID
          required: true
          schema:
            type: integer
        - name: body
          in: query
          description: \'Users\' object
          schema:
            type: object
      responses:
        "200":
          description: user fetched successfully

    delete:
      tags:
        - Users
      summary: deletes a user
      description: removes a user from the database
      parameters:
        - name: id
          in: path
          description: user ID
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: user deleted successfully
  /users/login/:
    post:
      tags:
        - Users
      summary: authenticates a user
      description: authenticates a user & gives them a JSON web token
      parameters:
        - name: body
          in: query
          description: object containing login credentials
          schema:
            type: object
      responses:
        "201":
          description: user created successfully
...